+++ raiden-dapp/tests/unit/views/select-hub-route.spec.ts
@@ -3,6 +3,7 @@ import Vuetify from 'vuetify';
 import { mount, Wrapper } from '@vue/test-utils';
 import store from '@/store';
 import { $identicon } from '../utils/mocks';
+import { connectAccount } from '../utils/store-utils';
 import SelectHubRoute from '@/views/SelectHubRoute.vue';
 
 Vue.use(Vuetify);
@@ -29,8 +30,7 @@ describe('SelectHubRoute.vue', () => {
   });
 
   test('connected displays actual route', async () => {
-    store.commit('account', '0x0000000000000000000000000000000000020001');
-    store.commit('loadComplete', true);
+    connectAccount();
     await wrapper.vm.$nextTick();
     expect(wrapper.find('home-stub').exists()).toBe(false);
     expect(wrapper.find('select-hub-stub').exists()).toBe(true);