+++ raiden-dapp/CHANGELOG.md
@@ -9,6 +9,7 @@
 - [#1786] Introduces snackbar display for notifications
 - [#1824] Listen to channel settle events and push notifications for them
 - [#2002] Add support to VUE_APP_MATRIX_LIST_URL transpile-time env var
+- [#1658] Add a disclaimer that the user needs to accept to get access to the app
 
 ### Changed
 
@@ -30,6 +31,7 @@
 [#1824]: https://github.com/raiden-network/light-client/issues/1824
 [#1875]: https://github.com/raiden-network/light-client/issues/1875
 [#2002]: https://github.com/raiden-network/light-client/issues/2002
+[#1658]: https://github.com/raiden-network/light-client/issues/1658
 
 ## [0.10.0] - 2020-07-13